test-automatiom/api-check/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
pyyaml>=6.0.1
jinja2>=3.1.3
apscheduler>=3.10.4
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
